Move local indexers into `indexers/local` folder
