+++ src/catalog.ts
@@ -7,13 +7,16 @@ import {
   DocumentListItem,
 } from "./document";
 import * as fs from "node:fs";
-import { JSONIndexer, JSONIndexerOpts } from "./indexers/json-indexer";
+import { JSONIndexer, JSONIndexerOpts } from "./indexers/local/json-indexer";
 import {
   DirectoryIndexer,
   DirectoryIndexerOpts,
-} from "./indexers/directory-indexer";
-import { TSVIndexer, TSVIndexerOpts } from "./indexers/tsv-indexer";
-import { ShopifyIndexer, ShopifyIndexerOpts } from "./indexers/shopify-indexer";
+} from "./indexers/local/directory-indexer";
+import { TSVIndexer, TSVIndexerOpts } from "./indexers/local/tsv-indexer";
+import {
+  ShopifyIndexer,
+  ShopifyIndexerOpts,
+} from "./indexers/local/shopify-indexer";
 
 export type CatalogConfig = {
   description: string;
