+++ src/index.ts
@@ -6,7 +6,7 @@ export * from "./cortex";
 export * from "./chat";
 export * from "./content";
 export * from "./document";
-export * from "./indexers/directory-indexer";
-export * from "./indexers/json-indexer";
-export * from "./indexers/tsv-indexer";
-export * from "./indexers/shopify-indexer";
+export * from "./indexers/local/directory-indexer";
+export * from "./indexers/local/json-indexer";
+export * from "./indexers/local/tsv-indexer";
+export * from "./indexers/local/shopify-indexer";
