+++ src/indexers/local/directory-indexer.ts
@@ -1,6 +1,6 @@
 import * as fs from "node:fs";
-import { Catalog } from "../catalog";
-import { FileDocument } from "../document";
+import { Catalog } from "../../catalog";
+import { FileDocument } from "../../document";
 
 export type DirectoryIndexerOpts = {
   rootDir: string;
