+++ src/indexers/local/json-indexer.ts
@@ -1,5 +1,5 @@
-import { Catalog } from "../catalog";
-import { JSONDocument } from "../document";
+import { Catalog } from "../../catalog";
+import { JSONDocument } from "../../document";
 
 // intentionally operates on any type
 /* eslint-disable  @typescript-eslint/no-explicit-any */
